+++ controller/gateway/controller.go
@@ -132,6 +132,10 @@ func (r *Reconciler) SetupWithManager(ctx context.Context, mgr ctrl.Manager) err
 			&gatewayv1.HTTPRoute{},
 			handler.EnqueueRequestsFromMapFunc(r.listGatewaysAttachedByHTTPRoute),
 			builder.WithPredicates(predicate.GenerationChangedPredicate{})).
+		Watches(
+			&gatewayv1.TLSRoute{},
+			handler.EnqueueRequestsFromMapFunc(r.listGatewaysAttachedByTLSRoute),
+		).
 		// watch Namespaces so that managed routes have correct status reflected in Gateway's
 		// status in status.listeners.attachedRoutes
 		// This is required to properly support Gateway's listeners.allowedRoutes.namespaces.selector.
@@ -189,6 +193,23 @@ func (r *Reconciler) SetupWithManager(ctx context.Context, mgr ctrl.Manager) err
 		)
 	} else {
 		log.Info(mgr.GetLogger(), "GRPCRoute CRD not found in cluster, skipping watch for GRPCRoute resources")
+	}
+
+	tcpRouteGVR := schema.GroupVersionResource{
+		Group:    gatewayv1.GroupName,
+		Version:  gatewayv1.GroupVersion.Version,
+		Resource: "tcproutes",
+	}
+	tcpRouteExist, err := crdChecker.CRDExists(tcpRouteGVR)
+	if err != nil {
+		return err
+	}
+	if tcpRouteExist {
+		blder.Watches(
+			&gatewayv1.TCPRoute{},
+			handler.EnqueueRequestsFromMapFunc(r.listGatewaysAttachedByTCPRoute),
+			builder.WithPredicates(predicate.GenerationChangedPredicate{}),
+		)
 	}
 
 	// Watch Secrets to requeue Gateways that reference them via listeners.tls.certificateRefs.