+++ controller/gateway/controller_watch.go
@@ -470,6 +470,44 @@ func (r *Reconciler) listGatewaysAttachedByGRPCRoute(ctx context.Context, obj cl
 	return listGatewaysAttachedByRoute(grpcRoute)
 }
 
+// listGatewaysAttachedByTCPRoute is a watch predicate which finds all Gateways mentioned
+// in TCPRoutes' ParentRefs.
+func (r *Reconciler) listGatewaysAttachedByTCPRoute(ctx context.Context, obj client.Object) []reconcile.Request {
+	logger := ctrllog.FromContext(ctx)
+
+	tcpRoute, ok := obj.(*gatewayv1.TCPRoute)
+	if !ok {
+		logger.Error(
+			fmt.Errorf("unexpected object type"),
+			"TCPRoute watch predicate received unexpected object type",
+			"expected", "*gatewayapi.TCPRoute", "found", reflect.TypeOf(obj),
+		)
+		return nil
+	}
+
+	gateways := &gatewayv1.GatewayList{}
+	if err := r.List(ctx, gateways); err != nil {
+		logger.Error(err, "Failed to list gateways in watch", "tcproute", client.ObjectKeyFromObject(tcpRoute))
+		return nil
+	}
+	var recs []reconcile.Request
+	for _, gateway := range gateways.Items {
+		for _, parentRef := range tcpRoute.Spec.ParentRefs {
+			if parentRef.Group != nil && string(*parentRef.Group) == gatewayv1.GroupName &&
+				parentRef.Kind != nil && string(*parentRef.Kind) == "Gateway" &&
+				string(parentRef.Name) == gateway.Name {
+				recs = append(recs, reconcile.Request{
+					NamespacedName: types.NamespacedName{
+						Namespace: gateway.Namespace,
+						Name:      gateway.Name,
+					},
+				})
+			}
+		}
+	}
+	return recs
+}
+
 // -----------------------------------------------------------------------------
 // GatewayReconciler - Config Defaults
 // -----------------------------------------------------------------------------