+++ test/helpers/generators.go
@@ -161,6 +161,43 @@ func GenerateHTTPRoute(namespace string, gatewayName, serviceName string, opts .
 	return httpRoute
 }
 
+// GenerateTCPRoute generates a TCPRoute to be used in tests.
+func GenerateTCPRoute(namespace string, gatewayName, serviceName string, portNumber int, opts ...func(*gatewayv1.TCPRoute)) *gatewayv1.TCPRoute {
+	tcpRoute := &gatewayv1.TCPRoute{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      uuid.NewString(),
+			Namespace: namespace,
+		},
+		Spec: gatewayv1.TCPRouteSpec{
+			CommonRouteSpec: gatewayv1.CommonRouteSpec{
+				ParentRefs: []gatewayv1.ParentReference{
+					{
+						Name: gatewayv1.ObjectName(gatewayName),
+					},
+				},
+			},
+			Rules: []gatewayv1.TCPRouteRule{
+				{
+					BackendRefs: []gatewayv1.BackendRef{
+						{
+							BackendObjectReference: gatewayv1.BackendObjectReference{
+								Name: gatewayv1.ObjectName(serviceName),
+								Port: new(gatewayv1.PortNumber(portNumber)),
+								Kind: new(gatewayv1.Kind("Service")),
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	for _, opt := range opts {
+		opt(tcpRoute)
+	}
+	return tcpRoute
+}
+
 // GenerateTLSRoute generates a TLSRoute to be used in tests.
 func GenerateTLSRoute(namespace string, gatewayName, serviceName string, portNumber int, opts ...func(*gatewayv1.TLSRoute)) *gatewayv1.TLSRoute {
 	tlsRoute := &gatewayv1.TLSRoute{